backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pandas~=2.2.3",
  "chromadb~=1.3.5",
  "dotenv~=0.9.9",
  "logging~=0.4.9.6",
  "uuid~=1.30",
  "sentence_transformers~=5.1.2",
  "torch>=2.2.0",
  "torchvision>=0.16.0",
  "groq>=0.11.0",
  "semantic-router>=0.1.12",
  "bytez>=2.0.4",
  "einops>=0.8.1",
  "fastapi[standard]~=0.121.2",
  "uvicorn[standard]",
  "pydantic-settings>=2.12.0",
  "sqlalchemy>=2.0.44",
  "asyncpg>=0.30.0",
  "psycopg2>=2.9.11",
]
